use js bigints

+++ src/lib.rs
@@ -4,14 +4,14 @@ use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 pub struct Factor {
-    prime: u32,
-    deg: u32,
+    prime: u64,
+    deg: usize,
 }
 
 #[wasm_bindgen]
 impl Factor {
     #[wasm_bindgen(constructor)]
-    pub fn new(prime: u32, deg: u32) -> Self {
+    pub fn new(prime: u64, deg: usize) -> Self {
         Self {
             prime: prime,
             deg: deg,
@@ -19,18 +19,18 @@ impl Factor {
     }
 
     #[wasm_bindgen(getter)]
-    pub fn prime(&self) -> u32 {
+    pub fn prime(&self) -> u64 {
         self.prime
     }
 
     #[wasm_bindgen(getter)]
-    pub fn deg(&self) -> u32 {
+    pub fn deg(&self) -> usize {
         self.deg
     }
 }
 
 #[wasm_bindgen]
-pub fn factors(n: u32) -> Vec<Factor> {
+pub fn factors(n: u64) -> Vec<Factor> {
     utils::factors(n)
         .into_iter()
         .map(|(k, v)| Factor::new(k, v))
